WIP 7 (put widget classes into separate files)
