+++ animestreamer/globals.py
@@ -1,0 +1,3 @@
+﻿from animestreamer.streamer import AnimeStreamer  # package install
+
+streamer = AnimeStreamer()
