+++ animestreamer/widgets/custom_footer.py
@@ -1,0 +1,31 @@
+﻿from rich.text import Text
+from textual.widgets import Footer
+
+
+class CustomFooter(Footer):
+
+    def make_key_text(self) -> Text:
+        """Override for styling and show only keys with description"""
+        text = Text(
+            style="white on rgb(98,98,98)",
+            no_wrap=True,
+            overflow="ellipsis",
+            justify="left",
+            end="",
+        )
+        for binding in self.app.bindings.shown_keys:
+            if not binding.description:
+                continue
+            key_display = (
+                binding.key.upper()
+                if binding.key_display is None
+                else binding.key_display
+            )
+            hovered = self.highlight_key == binding.key
+            key_text = Text.assemble(
+                (f" {key_display} ", "reverse" if hovered else "default on default"),
+                f" {binding.description} ",
+                meta={"@click": f"app.press('{binding.key}')", "key": binding.key},
+            )
+            text.append_text(key_text)
+        return text
